+++ src/main/java/com/jhogo/irsystem/car/CarDTO.java
@@ -1,8 +1,9 @@
 package com.jhogo.irsystem.car;
+import java.math.BigDecimal;
 public class CarDTO {
 
     private int manufactureYear,id;
-    private double saleValue, purchaseValue;
+    private BigDecimal saleValue, purchaseValue;
     private String model, brand, chassis;
     private boolean inStore;
     
@@ -22,19 +23,19 @@ public class CarDTO {
         this.manufactureYear = manufactureYear;
     }
     
-    public double getSaleValue() {
+    public BigDecimal getSaleValue() {
         return saleValue;
     }
     
-    public void setSaleValue(double saleValue) {
+    public void setSaleValue(BigDecimal saleValue) {
         this.saleValue = saleValue;
     }
     
-    public double getPurchaseValue() {
+    public BigDecimal getPurchaseValue() {
         return purchaseValue;
     }
     
-    public void setPurchaseValue(double purchaseValue) {
+    public void setPurchaseValue(BigDecimal purchaseValue) {
         this.purchaseValue = purchaseValue;
     }
